+++ lib/pubnub/events/channel_registration.rb
@@ -94,18 +94,6 @@ module Pubnub
       ].delete_if(&:blank?).join('/')
     end
 
-    def format_envelopes(response, request)
-      parsed_response, error = Formatter.parse_json(response.body)
-
-      error = response if parsed_response && response.code.to_i != 200
-
-      if error
-        error_envelope(parsed_response, error, request: request, response: response)
-      else
-        valid_envelope(parsed_response, request: request, response: response)
-      end
-    end
-
     def valid_envelope(parsed_response, req_res_objects)
       # {"status"=>200, "message"=>"OK", "service"=>"channel-registry", "error"=>false}
       Pubnub::Envelope.new(
@@ -140,29 +128,8 @@ module Pubnub
       )
     end
 
-    def error_envelope(parsed_response, error, req_res_objects)
-      Pubnub::ErrorEnvelope.new(
-          event: @event,
-          event_options: @given_options,
-          timetoken: nil,
-          status: {
-              code: req_res_objects[:response].code,
-              operation: define_operation,
-              client_request: req_res_objects[:request],
-              server_response: req_res_objects[:response],
-              data: nil,
-              category: (error ? Pubnub::Constants::STATUS_NON_JSON_RESPONSE : Pubnub::Constants::STATUS_ERROR),
-              error: true,
-              auto_retried: false,
-
-              current_timetoken: nil,
-              last_timetoken: nil,
-              subscribed_channels: nil,
-              subscribed_channel_groups: nil,
-
-              config: get_config
-          }
-      )
+    def current_operation
+      define_operation
     end
 
     def define_operation