+++ lib/pubnub/events/history.rb
@@ -12,6 +12,10 @@ module Pubnub
 
     private
 
+    def current_operation
+      Pubnub::Constants::OPERATION_HISTORY
+    end
+
     def timetoken(message)
       message['timetoken'] if @include_token
     end
@@ -49,18 +53,6 @@ module Pubnub
       params.merge!(reverse: 'true')       if @reverse
       params.merge!(include_token: 'true') if @include_token
       params
-    end
-
-    def format_envelopes(response, request)
-      parsed_response, error = Formatter.parse_json(response.body)
-
-      error = response if parsed_response && response.code.to_i != 200
-
-      if error
-        error_envelope(parsed_response, error, request: uri, response: response)
-      else
-        valid_envelope(parsed_response, request: request, response: response)
-                  end
     end
 
     def valid_envelope(parsed_response, req_res_objects)
@@ -107,31 +99,6 @@ module Pubnub
             end: finish,
             start: start
           }
-        }
-      )
-    end
-
-    def error_envelope(_parsed_response, error, req_res_objects)
-      Pubnub::ErrorEnvelope.new(
-        event: @event,
-        event_options: @given_options,
-        timetoken: nil,
-        status: {
-          code: req_res_objects[:response].code,
-          operation: Pubnub::Constants::OPERATION_HEARTBEAT,
-          client_request: req_res_objects[:request],
-          server_response: req_res_objects[:response],
-          data: nil,
-          category: (error ? Pubnub::Constants::STATUS_NON_JSON_RESPONSE : Pubnub::Constants::STATUS_ERROR),
-          error: true,
-          auto_retried: false,
-
-          current_timetoken: nil,
-          last_timetoken: nil,
-          subscribed_channels: nil,
-          subscribed_channel_groups: nil,
-
-          config: get_config
         }
       )
     end