+++ lib/pubnub/events/leave.rb
@@ -18,6 +18,10 @@ module Pubnub
 
     private
 
+    def current_operation
+      Pubnub::Constants::OPERATION_PRESENCE_LEAVE
+    end
+
     def path
       '/' + [
         'v2',
@@ -28,20 +32,6 @@ module Pubnub
         Formatter.channels_for_url(@channel),
         'leave'
       ].join('/')
-    end
-
-    def format_envelopes(response, request)
-      parsed_response, error = Formatter.parse_json(response.body)
-
-      error = response if parsed_response && response.code.to_i != 200
-
-      envelopes = if error
-                    error_envelope(parsed_response, error, request: request, response: response)
-                  else
-                    valid_envelope(parsed_response, request: request, response: response)
-                  end
-
-      envelopes
     end
 
     def valid_envelope(parsed_response, req_res_objects)
@@ -57,31 +47,6 @@ module Pubnub
           data: nil,
           category: Pubnub::Constants::STATUS_ACK,
           error: false,
-          auto_retried: false,
-
-          current_timetoken: nil,
-          last_timetoken: nil,
-          subscribed_channels: nil,
-          subscribed_channel_groups: nil,
-
-          config: get_config
-        }
-      )
-    end
-
-    def error_envelope(_parsed_response, error, req_res_objects)
-      Pubnub::ErrorEnvelope.new(
-        event: @event,
-        event_options: @given_options,
-        timetoken: nil,
-        status: {
-          code: req_res_objects[:response].code,
-          operation: Pubnub::Constants::OPERATION_HEARTBEAT,
-          client_request: req_res_objects[:request],
-          server_response: req_res_objects[:response],
-          data: nil,
-          category: (error ? Pubnub::Constants::STATUS_NON_JSON_RESPONSE : Pubnub::Constants::STATUS_ERROR),
-          error: true,
           auto_retried: false,
 
           current_timetoken: nil,