+++ lib/pubnub/events/publish.rb
@@ -39,6 +39,10 @@ module Pubnub
 
     private
 
+    def current_operation
+      Pubnub::Constants::OPERATION_PUBLISH
+    end
+
     def parameters
       @sequence_number        = sequence_number!
       @origination_time_token = @app.generate_ortt
@@ -91,18 +95,6 @@ module Pubnub
       nil
     end
 
-    def format_envelopes(response, request)
-      parsed_response, error = Formatter.parse_json(response.body)
-
-      error = response if parsed_response && response.code.to_i != 200
-
-      if error
-        error_envelope(parsed_response, error, request: request, response: response)
-      else
-        valid_envelope(parsed_response, request: request, response: response)
-                  end
-    end
-
     def valid_envelope(_parsed_response, req_res_objects)
       Pubnub::Envelope.new(
         event: @event,
@@ -116,31 +108,6 @@ module Pubnub
           data: nil,
           category: Pubnub::Constants::STATUS_ACK,
           error: false,
-          auto_retried: false,
-
-          current_timetoken: nil,
-          last_timetoken: nil,
-          subscribed_channels: nil,
-          subscribed_channel_groups: nil,
-
-          config: get_config
-        }
-      )
-    end
-
-    def error_envelope(_parsed_response, error, req_res_objects)
-      Pubnub::ErrorEnvelope.new(
-        event: @event,
-        event_options: @given_options,
-        timetoken: nil,
-        status: {
-          code: req_res_objects[:response].code,
-          operation: Pubnub::Constants::OPERATION_PUBLISH,
-          client_request: req_res_objects[:request],
-          server_response: req_res_objects[:response],
-          data: nil,
-          category: (error ? Pubnub::Constants::STATUS_NON_JSON_RESPONSE : Pubnub::Constants::STATUS_ERROR),
-          error: true,
           auto_retried: false,
 
           current_timetoken: nil,